Share

+++ src/main/java/com/ceojun7/wooricalendar/service/ShareService.java
@@ -6,7 +6,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.ceojun7.wooricalendar.model.CalendarEntity;
+import com.ceojun7.wooricalendar.model.ScheduleEntity;
 import com.ceojun7.wooricalendar.model.ShareEntity;
+import com.ceojun7.wooricalendar.persistence.CalendarRepository;
+import com.ceojun7.wooricalendar.persistence.ScheduleRepository;
 import com.ceojun7.wooricalendar.persistence.ShareRepository;
 
 import lombok.extern.slf4j.Slf4j;
@@ -31,6 +35,17 @@ public class ShareService {
 
   @Autowired
   private ShareRepository shareRepository;
+
+  @Autowired
+  private CalendarRepository calendarRepository;
+
+  @Autowired
+  private CalendarService calendarService;
+
+  @Autowired
+  private ScheduleRepository scheduleRepository;
+  @Autowired
+  private ScheduleService scheduleService;
 
   /**
    * methodName : create
@@ -64,12 +79,11 @@ public class ShareService {
     return shareRepository.findByShareNo(shareNo);
   }
 
-
   public List<ShareEntity> retrieveByEmail(final String email) {
     return shareRepository.findByMemberEntity_Email(email);
   }
 
-  public List<ShareEntity> retrieveByCalNo(final Long calNo){
+  public List<ShareEntity> retrieveByCalNo(final Long calNo) {
     return shareRepository.findByCalendarEntity_CalNo(calNo);
   }
 
@@ -111,9 +125,24 @@ public class ShareService {
    * @return shareNo
    * 
    */
-  public List<ShareEntity> delete(final ShareEntity shareEntity) {
-    shareRepository.delete(shareEntity);
-    Long calNo = shareEntity.getCalendarEntity().getCalNo();
+  public List<ShareEntity> delete(ShareEntity shareEntity) {
+    // Long calNo = shareEntity.getCalendarEntity().getCalNo(); // 구독했던 캘린더 번호 가져오기
+    List<ShareEntity> shareEntities = shareRepository.findByShareNo(shareEntity.getShareNo());
+    Long calNo = shareEntities.get(0).getCalendarEntity().getCalNo();
+
+    List<ShareEntity> entities = shareRepository.findByCalendarEntity_CalNo(calNo); // 캘린더 번호를 통하여 구독자 목록 가져오기
+    List<ScheduleEntity> scheduleEntities = scheduleRepository.findByCalendarEntity_CalNo(calNo); // 캘린더 번호를 통하여 일정가져오기
+
+    if (entities.size() == 1) { // 구독자가 한명이면(나만)
+      if (scheduleEntities.size() > 0) { // 일정이 하나 이상 존재하면
+        for (int i = 0; i < scheduleEntities.size(); i++) { // 모든 일정 탐색
+          scheduleRepository.delete(scheduleEntities.get(i)); // 모든 일정 삭제
+        }
+      }
+      shareRepository.delete(shareEntity); // 구독 취소
+      calendarService.delete(calendarRepository.findByCalNo(calNo).get(0)); // 캘린더 삭제
+    }
+    shareRepository.delete(shareEntity); // 구독자가 여러명일 경우 구독취소만 됨
 
     return shareRepository.findByShareNo(shareEntity.getShareNo());
   }
